XRDDEV-2138: API to return approved certification service details. (#1389)
* XRDDEV-2138: Api to get certification service details

* XRDDEV-2138: Introducing new type ApprovedCertificationServiceListItem for returning list of approved certification services.
Modifying ApprovedCertificationService openapi type removing ocspResponders, intermediateCas and leaving only required to display properties.

* XRDDEV-2138: Deleting created but unused dtos.

* XRDDEV-2138: Improving controller to use dto instead of entity when saving approved certification service

* XRDDEV-2138 Using converter, entity implementations instead of mock in unit test.
Using static imports.

+++ src/centralserver/admin-service/core/src/main/java/org/niis/xroad/centralserver/restapi/repository/ApprovedCaRepository.java
@@ -30,6 +30,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ApprovedCaRepository extends JpaRepository<ApprovedCa, Long> {
+public interface ApprovedCaRepository extends JpaRepository<ApprovedCa, Integer> {
 
 }
